# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.1.0",
    "title": "DataPorts API with swagger",
    "description": "This is a simple REST API made with Node.JS and documented with Swagger",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "José A. Clemente",
      "email": "jclemente@prodevelop.es"
    }
  },
  "basePath": "/dac",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/on_demand": {
      "post": {
        "tags": [
          "On-demand"
        ],
        "summary": "Create an on demand agent",
        "description": "Method called from Semantic & Interoperability Component to create an On-demand agent",
        "operationId": "createOnDemand",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Create an on demand agent",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OnDemand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/images": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Returns the images availables on this host",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/images/{id}/config_datasource": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Returns a DataSource object with information about an image",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to retrieve its datasource",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/images/{id}/template": {
      "get": {
        "tags": [
          "Images"
        ],
        "operationId": "getTemplate",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "bearerToken": []
          }
        ],
        "summary": "Returns a template for creating an agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to retrieve its template",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/images/{id}": {
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Deletes an image",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the image that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
    "/notification": {
      "get": {
        "tags": [
          "Notification"
        ],
        "summary": "Returns all the notifications available in the system",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of elements to return in the query. This number is limited to a max of 100 elements",
            "type": "number"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of elemnts skipped from thq beginning of the query",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            },
            "headers": {
              "x-total-count": {
                "description": "Total number of elements available",
                "type": "number"
              }
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
    "/notification/{id}": {
      "get": {
        "tags": [
          "Notification"
        ],
        "summary": "Returns a specific notification",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the notification that we want to retrieve",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notification"
        ],
        "summary": "Deletes a Notification",
        "security": [
          {
            "bearerToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the notification that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succesful operation",
            "schema": {
              "$ref": "#/definitions/GenericAnswer"
            }
          },
          "404": {
            "description": "Generic error",
            "schema": {
              "$ref": "#/definitions/GenericMessage"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token with the `Bearer` prefix, E.g., \"Bearer easkkassjejm.dadsaeea.ssdaee\""
    }
  },
  "definitions": {
    "OnDemand": {
      "type": "object",
      "required": [
        "name",
        "image",
        "environment"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "environment": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "key": {
                "type": "string"
              },
              "value": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "GenericMessage": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "GenericAnswer": {
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}